requirements.txt
playwright
reflex==0.8.17
pandas
openpyxl
python-dotenv